Finish up moving pivot role permissions from core to modules
