+++ backend/dataall/modules/dataset_sharing/cdk/__init__.py
@@ -1,0 +1,3 @@
+from dataall.modules.dataset_sharing.cdk import pivot_role_data_sharing_policy
+
+__all__ = ["pivot_role_data_sharing_policy"]
