+++ backend/dataall/modules/mlstudio/cdk/pivot_role_mlstudio_policy.py
@@ -17,16 +17,11 @@ class MLStudioPivotRole(PivotRoleStatementSet):
                 actions=[
                     'sagemaker:ListTags',
                     'sagemaker:DescribeUserProfile',
-                    'sagemaker:StopNotebookInstance',
-                    'sagemaker:CreatePresignedNotebookInstanceUrl',
-                    'sagemaker:DescribeNotebookInstance',
-                    'sagemaker:StartNotebookInstance',
                     'sagemaker:AddTags',
                     'sagemaker:DescribeDomain',
                     'sagemaker:CreatePresignedDomainUrl',
                 ],
                 resources=[
-                    f'arn:aws:sagemaker:*:{self.account}:notebook-instance/{self.env_resource_prefix}*',
                     f'arn:aws:sagemaker:*:{self.account}:domain/*',
                     f'arn:aws:sagemaker:*:{self.account}:user-profile/*/*',
                 ],
@@ -35,10 +30,21 @@ class MLStudioPivotRole(PivotRoleStatementSet):
                 sid='SageMakerNotebookInstances',
                 effect=iam.Effect.ALLOW,
                 actions=[
-                    'sagemaker:ListNotebookInstances',
                     'sagemaker:ListDomains',
                     'sagemaker:ListApps',
                     'sagemaker:DeleteApp',
+                ],
+                resources=['*'],
+            ),
+            iam.PolicyStatement(
+                sid='EC2SGMLStudio',
+                effect=iam.Effect.ALLOW,
+                actions=[
+                    'ec2:DescribeSubnets',
+                    'ec2:DescribeSecurityGroups',
+                    'ec2:DescribeVpcs',
+                    'ec2:DescribeInstances',
+                    'ec2:DescribeNetworkInterfaces',
                 ],
                 resources=['*'],
             ),
