+++ backend/dataall/modules/worksheets/cdk/__init__.py
@@ -1,0 +1,3 @@
+from dataall.modules.worksheets.cdk import pivot_role_worksheets_policy
+
+__all__ = ["pivot_role_worksheets_policy"]
