+++ backend/dataall/modules/worksheets/cdk/pivot_role_worksheets_policy.py
@@ -2,9 +2,9 @@ from dataall.core.environment.cdk.pivot_role_stack import PivotRoleStatementSet
 from aws_cdk import aws_iam as iam
 
 
-class AthenaPivotRole(PivotRoleStatementSet):
+class WorksheetsPivotRole(PivotRoleStatementSet):
     """
-    Class including all permissions needed  by the pivot role to work with AWS Athena.
+    Class including all permissions needed  by the pivot role to work with Athena
     It allows pivot role to:
     - ....
     """
